refactored rendering and submitting of action details so we can make for better code reuse

+++ rah/views.py
@@ -70,48 +70,35 @@ def action_browse(request):
     cats = ActionCat.objects.all()
     return render_to_response('rah/action_browse.html', {'cats':cats}, context_instance=RequestContext(request))
 
-def action_cat(request, catSlug):
+def action_cat(request, cat_slug):
     """View an action category page with links to actions in that category"""
-    cat     = get_object_or_404(ActionCat, slug=catSlug)
+    cat     = get_object_or_404(ActionCat, slug=cat_slug)
     actions = Action.objects.filter(category=cat.id)
         
     return render_to_response('rah/action_cat.html', {'cat': cat, 'actions': actions}, 
                                 context_instance=RequestContext(request))
 
-def action_detail(request, catSlug, actionSlug):
+def action_detail(request, cat_slug, action_slug):
     """Detail page for an action"""
     # Lookup the action
-    action = get_object_or_404(Action, slug=actionSlug)
-    
-    #  Handle the POST if a task is being completed and task_id is an integer
-    if request.method == 'POST' and request.POST.get('task_id').isdigit():
-        try:
-            # Look for a row in UserActionTask table and delete it if task_completed is unchecked
-            user_action_task = UserActionTask.objects.get(user=request.user, action_task=request.POST.get('task_id'))
-            if request.POST.get('task_completed') == None:
-                user_action_task.delete()
-        except:
-            # There was no row in UserActionTask, so create one if task_completed is checked
-            if request.POST.get('task_completed') == 'on':
-                UserActionTask(user=request.user, action_task=ActionTask(pk=request.POST.get('task_id'))).save()
-    
-    # Get action tasks conditionally depending on whether or not they're logged in
-    if(request.user.is_authenticated()):
-        # If they're logged in, grab the completed time for any completed tasks
-        action_tasks = ActionTask.objects.filter(action=action.id).extra(
-            select_params = (request.user.id,), 
-            select = { 'completed': 'SELECT rah_useractiontask.completed \
-                                     FROM rah_useractiontask \
-                                     WHERE rah_useractiontask.user_id = %s AND \
-                                     rah_useractiontask.action_task_id = rah_actiontask.id' }
-        )
-    else:
-        action_tasks = ActionTask.objects.filter(action=action.id)
+    action = get_object_or_404(Action, slug=action_slug)
+    action_tasks = ActionTask.get_action_tasks_by_action_optional_user(action, request.user)
     
     return render_to_response('rah/action_detail.html', {
                                 'action': action,
                                 'action_tasks': action_tasks
                               }, context_instance=RequestContext(request))
+                              
+def action_task(request, action_task_id):
+    #  Handle the POST if a task is being completed and task_id is an integer
+    action_task = get_object_or_404(ActionTask, id=action_task_id)
+    if request.method == 'POST' and request.user.is_authenticated(): #and request.POST.get('task_id').isdigit():
+        user_action_task, created = UserActionTask.objects.get_or_create(user=request.user, action_task=action_task)
+        print "HAS BEEN CREATED: %s" % (created)
+        print "IS TASK COMPLETE: %s" % (request.POST.get('task_completed'))
+        if request.POST.get('task_completed') == None:
+            user_action_task.delete()
+    return redirect('www.rah.views.action_detail', cat_slug=action_task.action.category.slug, action_slug=action_task.action.slug)
 
 def profile(request, username):
     """docstring for profile"""
